+++ NMSIS/Core/Include/core_feature_base.h
@@ -1747,14 +1747,13 @@ __STATIC_FORCEINLINE void __set_hpm_counter(unsigned long idx, uint64_t value)
 }
 
 /**
- * \brief   Get value of selected high performance monitor couner
+ * \brief   Get value of selected high performance monitor counter
  * \param [in]    idx     HPMCOUNTERx CSR index(3-31)
- * \param [in]    event   HPMCOUNTERx Register value to set
  * \details
  * Get high performance monitor counter register value
  * \return               HPMCOUNTERx Register value
  */
-__STATIC_FORCEINLINE unsigned long __get_hpm_counter(unsigned long idx)
+__STATIC_FORCEINLINE uint64_t __get_hpm_counter(unsigned long idx)
 {
 #if __RISCV_XLEN == 32
     volatile uint32_t high0, low, high;
@@ -1946,6 +1945,52 @@ __STATIC_FORCEINLINE unsigned long __get_hpm_counter(unsigned long idx)
 #else
     switch (idx) {
 #endif
+        default: return 0;
+    }
+}
+
+/**
+ * \brief   Get value of selected high performance monitor counter
+ * \param [in]    idx     HPMCOUNTERx CSR index(3-31)
+ * \details
+ * Get high performance monitor counter register value without high
+ *          32 bits when XLEN=32
+ * \return               HPMCOUNTERx Register value
+ */
+__STATIC_FORCEINLINE unsigned long __read_hpm_counter(unsigned long idx)
+{
+    switch (idx) {
+        case 0: return __read_cycle_csr();
+        case 2: return __read_instret_csr();
+        case 3: return __RV_CSR_READ(CSR_MHPMCOUNTER3);
+        case 4: return __RV_CSR_READ(CSR_MHPMCOUNTER4);
+        case 5: return __RV_CSR_READ(CSR_MHPMCOUNTER5);
+        case 6: return __RV_CSR_READ(CSR_MHPMCOUNTER6);
+        case 7: return __RV_CSR_READ(CSR_MHPMCOUNTER7);
+        case 8: return __RV_CSR_READ(CSR_MHPMCOUNTER8);
+        case 9: return __RV_CSR_READ(CSR_MHPMCOUNTER9);
+        case 10: return __RV_CSR_READ(CSR_MHPMCOUNTER10);
+        case 11: return __RV_CSR_READ(CSR_MHPMCOUNTER11);
+        case 12: return __RV_CSR_READ(CSR_MHPMCOUNTER12);
+        case 13: return __RV_CSR_READ(CSR_MHPMCOUNTER13);
+        case 14: return __RV_CSR_READ(CSR_MHPMCOUNTER14);
+        case 15: return __RV_CSR_READ(CSR_MHPMCOUNTER15);
+        case 16: return __RV_CSR_READ(CSR_MHPMCOUNTER16);
+        case 17: return __RV_CSR_READ(CSR_MHPMCOUNTER17);
+        case 18: return __RV_CSR_READ(CSR_MHPMCOUNTER18);
+        case 19: return __RV_CSR_READ(CSR_MHPMCOUNTER19);
+        case 20: return __RV_CSR_READ(CSR_MHPMCOUNTER20);
+        case 21: return __RV_CSR_READ(CSR_MHPMCOUNTER21);
+        case 22: return __RV_CSR_READ(CSR_MHPMCOUNTER22);
+        case 23: return __RV_CSR_READ(CSR_MHPMCOUNTER23);
+        case 24: return __RV_CSR_READ(CSR_MHPMCOUNTER24);
+        case 25: return __RV_CSR_READ(CSR_MHPMCOUNTER25);
+        case 26: return __RV_CSR_READ(CSR_MHPMCOUNTER26);
+        case 27: return __RV_CSR_READ(CSR_MHPMCOUNTER27);
+        case 28: return __RV_CSR_READ(CSR_MHPMCOUNTER28);
+        case 29: return __RV_CSR_READ(CSR_MHPMCOUNTER29);
+        case 30: return __RV_CSR_READ(CSR_MHPMCOUNTER30);
+        case 31: return __RV_CSR_READ(CSR_MHPMCOUNTER31);
         default: return 0;
     }
 }